refactor: types/animationTypes => types/animation
