+++ types/animation.ts
